refactor: #32; move helper sortNumbersAr() into utils file
